feat: add functional(закрыл доступ к боту)

+++ bot/handlers.py
@@ -1,9 +1,9 @@
 #!/usr/bin/env python3
 from aiogram import Bot, Dispatcher, executor, types
 from aiogram.types import ReplyKeyboardRemove
-from bot.screpbot import function_screp, screp_iter, lst
+from bot.screpers import function_screp, screp_iter, lst
 from bot.weather import get_weather
-from bot.tkn import token_bot
+from bot.tkn import token_bot, USER_ID
 from bot.keyboardd import kb, kb2, kbf, kbw, kbtr, cancelb, kbrecord, kbday
 from bot.baza import add_tren, get_workout_record, get_workout_all_record, update_tren, get_rowid
 from datetime import date
@@ -19,6 +19,7 @@ def main():
     storage = MemoryStorage()
     bot = Bot(token = token_bot)
     dp = Dispatcher(bot, storage=storage)
+     
 
 # _____________________________________________________________________________________________________________
     
@@ -29,21 +30,27 @@ def main():
     @dp.message_handler(commands=['start', 'старт'])
     @dp.message_handler(Text(equals=['старт','start','Старт','Start']))
     async def send_welcome(message: types.Message):
-        await message.reply(f'Привет, {message.from_user.first_name} \U0001F464, я бот секретарь! \u270D\n\
+        if message.from_user.id == USER_ID:
+            await message.reply(f'Привет, {message.from_user.first_name} \U0001F464, я бот секретарь! \u270D\n\
     жми /help или воспользуйся клавиатурой,\
     чтобы подать мне команду!', reply_markup = kb)
+        else:
+            await message.reply('У Вас нет доступа!!!')
 
-
+    
     # отработка команды help
-    @dp.message_handler(commands=['help', 'отмена'])
+    @dp.message_handler(commands=['help'])
     @dp.message_handler(Text(equals=['помощь','Помощь','Help','help','back','Back','назад','Назад']))
     async def send_help(message: types.Message):
-        await message.reply('\u2193  Доступные команды бота: \u2193\n\n\
+        if message.from_user.id == USER_ID:
+            await message.reply('\u2193  Доступные команды бота: \u2193\n\n\
     /fonda - информация о фондовом рынке\n\n\
     /crypto - информация по криптовалютам\n\n\
     /weath - информация о погоде\n\n\
     /infotren -  информация о ведении журнала тренировок', reply_markup = kb)
-        await message.delete()
+            await message.delete()
+        else:
+            await message.reply('У Вас нет доступа!!!')
 # _____________________________________________________________________________________________________________
 
 
@@ -53,22 +60,31 @@ def main():
     @dp.message_handler(commands=['fonda'])
     @dp.message_handler(Text(equals=['Фонда','фонда','fonda','Fonda']))
     async def get_keyboard_info_fonda(message: types.Message):
-        await message.reply('Выбери нужный раздел', reply_markup = kbf)
+        if message.from_user.id == USER_ID:
+            await message.reply('Выбери нужный раздел', reply_markup = kbf)
+        else:
+            await message.reply('У Вас нет доступа!!!')
     
     # отработка комаманды 'crypto'
     @dp.message_handler(commands=['crypto'])
     @dp.message_handler(Text(equals=['Крипта','крипта','Crypto','crypto']))
     async def send_crypto(message: types.Message):
-        await message.reply('Функционал отсутствует...', reply_markup = kbf)
+        if message.from_user.id == USER_ID:
+            await message.reply('Функционал отсутствует...', reply_markup = kbf)
+        else:
+            await message.reply('У Вас нет доступа!!!')
 
 
     # отработка команды market, отдает текущие цены
     @dp.message_handler(commands=['market'])
     @dp.message_handler(Text(equals=['market','Market','Рынок','рынок']))
     async def send_fonda(message: types.Message):
-        await message.answer('Жди, собираю информацию... \u23F3')
-        await message.reply(screp_iter(function_screp(lst)))
-        await message.delete()
+        if message.from_user.id == USER_ID:
+            await message.answer('Жди, собираю информацию... \u23F3')
+            await message.reply(screp_iter(function_screp(lst)))
+            await message.delete()
+        else:
+            await message.reply('У Вас нет доступа!!!')
     
 
     # ....................................информация о компаниях и токенах.........................................
@@ -77,27 +93,39 @@ def main():
     @dp.message_handler(commands=['info'])
     @dp.message_handler(Text(equals=['Инфа','инфа','Инфо','инфо']))
     async def info_message(message : types.Message):
-        message_text =f'Блок информации об представленых компаниях'
-        await message.answer(message_text,reply_markup=kb2)
+        if message.from_user.id == USER_ID:
+            message_text =f'Блок информации об представленых компаниях'
+            await message.answer(message_text,reply_markup=kb2)
+        else:
+            await message.reply('У Вас нет доступа!!!')
     
 
     #  инфо-заглушки
     @dp.message_handler(commands=['BYND'])
     async def info_BYND(message: types.Message):
-        await message.reply('Beyond meat\nКомпания специализирующаяся на производстве растительного мяса.\n\
+        if message.from_user.id == USER_ID:
+            await message.reply('Beyond meat\nКомпания специализирующаяся на производстве растительного мяса.\n\
     В качестве сырья используется горох.\nВ 2019 компания вышла  на IPO с 25$ за акцию.\n\
     На территории США продукция Beyon meat используется в продуктах McDonald\'s и Burger King', reply_markup=ReplyKeyboardRemove())
+        else:
+            await message.reply('У Вас нет доступа!!!')
 
     @dp.message_handler(commands=['EXEL'])
     async def info_EXEL(message: types.Message):
-        await message.reply('Exelixis, Inc\nБиотех, специализирующийся на разработке лекарственых\n\
+        if message.from_user.id == USER_ID:
+            await message.reply('Exelixis, Inc\nБиотех, специализирующийся на разработке лекарственых\n\
     препаратов от различных форм рака.\nОсновная молекула компании - кабозантиниб.\n\
     Сотрудничает со многими фармацептическими "монстрами".', reply_markup=ReplyKeyboardRemove())
+        else:
+            await message.reply('У Вас нет доступа!!!')
 
     @dp.message_handler(commands=['MU'])
     async def info_MU(message: types.Message):
-        await message.reply('Micron Technology, Inc\nИзготавливает полуппроводники(входит в ТОП 5 компаний)\
+        if message.from_user.id == USER_ID:
+            await message.reply('Micron Technology, Inc\nИзготавливает полуппроводники(входит в ТОП 5 компаний)\
     , основная часть которых - различные виды памяти.', reply_markup=ReplyKeyboardRemove())
+        else:
+            await message.reply('У Вас нет доступа!!!')
 # _____________________________________________________________________________________________________________
 
 
@@ -107,19 +135,25 @@ def main():
     @dp.message_handler(commands=['weath'])
     @dp.message_handler(Text(equals=['погода','Погода','Weath','weatch']))
     async def send_weath(message: types.Message):
-        await message.reply('Отправьте свои координаты для определения погоды в вашем районе', reply_markup=kbw)
-        @dp.message_handler(content_types=['location'])
-        async def location(message: types.Message):
-            lat = message.location.latitude
-            lon = message.location.longitude
-            await message.reply(f'широта:{lat}, долгота:{lon}')
-            await message.reply(get_weather(lat, lon), reply_markup=ReplyKeyboardRemove())
+        if message.from_user.id == USER_ID:
+            await message.reply('Отправьте свои координаты для определения погоды в вашем районе', reply_markup=kbw)
+            @dp.message_handler(content_types=['location'])
+            async def location(message: types.Message):
+                lat = message.location.latitude
+                lon = message.location.longitude
+                await message.reply(f'широта:{lat}, долгота:{lon}')
+                await message.reply(get_weather(lat, lon), reply_markup=ReplyKeyboardRemove())
+        else:
+            await message.reply('У Вас нет доступа!!!')
 
     # получение текущей погоды в Липецке
     @dp.message_handler(commands=['Липецк'])
     @dp.message_handler(Text(equals=['Липецк погода','Липецк']))
     async def send_weath(message: types.Message):
-        await message.reply(get_weather(),reply_markup=ReplyKeyboardRemove())
+        if message.from_user.id == USER_ID:
+            await message.reply(get_weather(),reply_markup=ReplyKeyboardRemove())
+        else:
+            await message.reply('У Вас нет доступа!!!')
 # _____________________________________________________________________________________________________________
 
 
@@ -137,9 +171,12 @@ def main():
     @dp.message_handler(commands=['tren','трен','Трен'])
     @dp.message_handler(Text(equals=['добавить тренировку','Добавить тренировку']))
     async def gt_tren(message : types.Message):
-        await Tren().bic.set()
-        await message.reply('Введи кол-во повторов на бицепс:\nЕсли повторов нет введи:  -\nОтмена - чтобы прекратить заполнение и выйти.', reply_markup=cancelb)
-    
+        if message.from_user.id == USER_ID:
+            await Tren().bic.set()
+            await message.reply('Введи кол-во повторов на бицепс:\nЕсли повторов нет введи:  -\nОтмена - чтобы прекратить заполнение и выйти.', reply_markup=cancelb)
+        else:
+            await message.reply('У Вас нет доступа!!!')
+
     # Добавляем возможность отмены, если пользователь передумал заполнять
     @dp.message_handler(state='*', commands='отмена')
     @dp.message_handler(Text(equals='отмена', ignore_case=True), state='*')
@@ -201,8 +238,11 @@ def main():
     @dp.message_handler(commands=['needed'])
     @dp.message_handler(Text(equals=['Получить запись журнала','получить запись','получить запись журнала', 'нужна запись','запись журнала','Запись журнала']))
     async def gt_tren(message : types.Message):
-        await Record.tabl.set()
-        await message.reply('Введи название стобца: \n\nthe_date - дата\nday - день недели\nbiceps - бицепс\nwaist - пояс\nchest - грудь\ntriceps - трицепс',reply_markup=kbrecord)
+        if message.from_user.id == USER_ID:
+            await Record.tabl.set()
+            await message.reply('Введи название стобца: \n\nthe_date - дата\nday - день недели\nbiceps - бицепс\nwaist - пояс\nchest - грудь\ntriceps - трицепс',reply_markup=kbrecord)
+        else:
+            await message.reply('У Вас нет доступа!!!')
 
     @dp.message_handler(state=Record.tabl)
     async def process_biceps(message: types.Message, state: FSMContext):
@@ -237,8 +277,11 @@ def main():
     @dp.message_handler(commands=['journal'])
     @dp.message_handler(Text(equals=['задать лимит записей','получить лимит записей','получить лимит', 'нужен лимит записей','лимит журнала']))
     async def gt_jornal_limit(message : types.Message):
-        await Journal.item.set()
-        await message.answer('Введи задай нужное кол-во записей для вывода:')
+        if message.from_user.id == USER_ID:
+            await Journal.item.set()
+            await message.answer('Введи задай нужное кол-во записей для вывода:')
+        else:
+            await message.reply('У Вас нет доступа!!!')
     
     @dp.message_handler(state=Journal.item)
     async def process_journal(message: types.Message, state: FSMContext):
@@ -253,7 +296,10 @@ def main():
     @dp.message_handler(commands=['журнал'])
     @dp.message_handler(Text(equals=['журнал','Журнал']))
     async def gt_journal(message : types.Message):
-        await message.answer(get_workout_all_record(7))
+        if message.from_user.id == USER_ID:
+            await message.answer(get_workout_all_record(7))
+        else:
+            await message.reply('У Вас нет доступа!!!')
     
     # получение уникального айди
 
@@ -264,8 +310,11 @@ def main():
     @dp.message_handler(commands=['rowid'])
     @dp.message_handler(Text(equals=['получить ровид ID','получить айди', 'получить powid id']))
     async def gt_rowid(message : types.Message):
-        await Rowid.value.set()
-        await message.answer('Задай дату записи в формате: "2022-07-14"\n\nКАВЫЧКИ ОБЯЗАТЕЛЬНЫ! ')
+        if message.from_user.id == USER_ID:
+            await Rowid.value.set()
+            await message.answer('Задай дату записи в формате: "2022-07-14"\n\nКАВЫЧКИ ОБЯЗАТЕЛЬНЫ! ')
+        else:
+            await message.reply('У Вас нет доступа!!!')
     
     @dp.message_handler(state=Rowid.value)
     async def process_rowid(message: types.Message, state: FSMContext):
@@ -289,14 +338,12 @@ def main():
     @dp.message_handler(commands=['update'])
     @dp.message_handler(Text(equals=['редактировать журнал']))
     async def update_journal(message : types.Message):
-        await Update_journal.name_column.set()
-        await message.answer('Введи название стобца: \n\nthe_date - дата\nday - день недели\nbiceps - бицепс\nwaist - пояс\nchest - грудь\ntriceps - трицепс',reply_markup=kbrecord)
-    #async def gt_tren(message : types.Message):
-        #name_column = message.text.split(',')[0][8:]
-        #new_value = message.text.split(',')[1]
-        #rowid = message.text.split(',')[2]
-       # await message.answer(update_tren(name_column, new_value, rowid))
-    
+        if message.from_user.id == USER_ID:
+            await Update_journal.name_column.set()
+            await message.answer('Введи название стобца: \n\nday - день недели\nbiceps - бицепс\nwaist - пояс\nchest - грудь\ntriceps - трицепс',reply_markup=kbrecord)
+        else:
+            await message.reply('У Вас нет доступа!!!')
+
     @dp.message_handler(state=Update_journal.name_column)
     async def process_rowid(message: types.Message, state: FSMContext):
         async with state.proxy() as data_state:
@@ -319,7 +366,7 @@ def main():
         name_column = data_state['name_column']
         new_value = data_state['new_value']
         rowid = data_state['rowid']
-        await message.answer(update_tren(name_column, new_value, rowid),reply_markup=ReplyKeyboardRemove())
+        await message.reply(update_tren(name_column, new_value, rowid))
 
 
     #........................................информация о работе с журналом........................
@@ -328,35 +375,19 @@ def main():
     @dp.message_handler(commands=['infotren'])
     @dp.message_handler(Text(equals=['инфо о журнале']))
     async def gt_tren(message : types.Message):
-        await message.answer('\
-ПОСМОТРЕТЬ ЗАПИСЬ В ЖУРНАЛЕ ТРЕНИРОВОК:\n\n\
-:/needed имя столбца,кол-во повторов\n\
-Пример:/needed biceps,23\n\
-Пример выдает журнал тренировок  на бицепс с 23 повторениями.\n\
-Имена таблиц:\n\
-        the_data    дата\n        day         день\n        biceps   бицепс\n        waist      пояс\n        chest     грудь\n        triceps  трицепс\n\
-ДЕНЬ НЕДЕЛИ УКАЗЫВАТЬ В \' КАВЫЧКАХ\' -- "вторник"\n\
-\n\n\
-ПОСМОТРЕТЬ ЖУРНАЛ ТРЕНИРОВОК\n\
-Лимиритрованый запрос журнала:\n\
-Команда: /journal *, где * - кол-во записей;отдает заданое кол-во записей\n\
-Пример: /jourrnal 2\n\n\
+        if message.from_user.id == USER_ID:
+            await message.answer('\
+КОМАНДА журнал :\nПоказывает последние 7 записей.\n\
 ___________________________________________\n\n\
-ДОБАВЛЕНИЕ ЗАПИСИ В ЖУРНАЛ:\n\n\
-: /tren бицепс,пояс,грудь,трицепс\n\
-Пример: /tren 23,15,20,12\n\
-При успешном добавлении,будет отображено соответствующее сообщение!\n\
+КОМАНДА добавить тренировку:\nдобавляет новую запись.\n\
 ___________________________________________\n\n\
-ПОЛУЧЕНИЕ УНИКАЛЬНОГО ID ЗАПИСИ\n\n\
-Команда rowid отдает "rowid id" по дате\n\
-Синтаксис: /rowid "2022-07-15"\n\
-КАВЫЧКИ ОБЯЗАТЕЛЬНЫ\n\
+КОМАНДА получить лимит записей:\nПоказывает заданное кол-во записей от начала ведения журнала.\n\
 ___________________________________________\n\n\
-РЕДАКТИРОВАНИЕ ЖУРНАЛА\n\n\
-Команда update - вносит изменения в таблицу\n\
-: /update столбец новое-значение rowid_ID\n\
-Синтаксис: /update day, "Среда", 3\n\
-КАВЫЧКИ ОБЯЗАТЕЛЬНЫ',reply_markup = kbtr)
+КОМАНДА получить запись:\nВыдает нужную запись.\n\
+___________________________________________\n\n\
+КОМАНДА редактировать журнал: Редактирование нужной записи.',reply_markup = kbtr)
+        else:
+            await message.reply('У Вас нет доступа!!!')
 
 # _____________________________________________________________________________________________________________
